+++ CLI/src/Musical Structures/Note/Pitch.h
@@ -11,19 +11,19 @@ namespace ReMu {
 	protected:
 		char step;
 		Accidental accidental = None;
-		float duration = -1;
+		int duration = -1;
 	public:
 		Pitch() { }
 		Pitch(char __note, Accidental __accidental = None) : step(__note), accidental(__accidental) {  }
-		Pitch(char __note, float __duration, Accidental __accidental = None) : step(__note), duration(__duration), accidental(__accidental) { }
+		Pitch(char __note, int __duration, Accidental __accidental = None) : step(__note), duration(__duration), accidental(__accidental) { }
 
 		inline void setStep(char step) { this->step = step; }
 		inline void setAccidental(Accidental modifier) { this->accidental = modifier; }
-		inline void setDuration(float duration) { this->duration = duration; }
+		inline void setDuration(int duration) { this->duration = duration; }
 
 		inline char getStep() { return step; }
 		inline Accidental getAccidental() { return accidental; }
-		inline float getDuration() { return duration; }
+		inline int getDuration() { return duration; }
 
 		friend bool operator==(const Pitch& lhs, const Pitch& rhs);
 		friend std::ostream& operator<<(std::ostream& os, const Pitch& note);
